requirements.txt
accelerate==0.16.0
datasets==2.9.0
editdistance==0.6.2
einops==0.6.0
# flax==0.6.7
huggingface_hub==0.12.0
importlib_metadata==6.0.0
ipdb==0.13.11
# jax==0.4.7
matplotlib==3.6.3
modelcards==0.1.6
msgpack_python==0.5.6
numpy==1.24.1
omegaconf==2.3.0
onnxruntime==1.14.1
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
pytorch_lightning==1.9.0
PyYAML==6.0
requests==2.28.2
scipy==1.10.1
torch==1.13.1
torchvision==0.14.1
tqdm==4.64.1
transformers==4.26.0
xformers==0.0.16
tokenizers==0.13.2
torch-tb-profiler==0.4.1
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6
pyarrow==11.0.0

# 自分で追加
shapely==2.0.6
opencv-python==4.10.0.84
hydra-core==1.3.2
